backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
pypng==0.20220715.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
